Housing-prediction/requirements.txt
Flask
gunicorn
sklearn
pandas
PyYAML
-e .